extract transformSpriteData helper
